+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 #### 功能构建
 
+- 定义实体的操作服务。
+  - com.dwarfeng.scg.stack.service.NodeVariableOperateService。
+
 - 定义实体及其维护服务，并通过单元测试。
   - com.dwarfeng.scg.stack.bean.entity.NodeVariable。
   - com.dwarfeng.scg.stack.bean.entity.CommonVariable。